Address review feedback on TwinCAT method parsing
- Move the TwinCAT parser tests into a sibling tests module. The parser
  module had grown past the project's 1000-line limit; it is now 494
  lines with the tests in twincat_parser/tests.rs. The XML fixtures are
  moved verbatim.
- Drop the plan reference from the method_declaration comment.
- Drop the issue references from the new comments.
